fix: Pointer argument is used instead of global variable
