+++ tests/evaluation_test.cpp
@@ -8,11 +8,12 @@ protected:
     void SetUp() override {
         // Common setup code that runs before each test case
         board_init();
-        evaluate_init();
+        //evaluate_init();
     }
 };
 
 TEST_F(EvaluationTest, d3c5) {
+    ArrStruct a = make_score();
     board black_d3c5;
     black_d3c5.translate_from_arr(TestBoard::d3c5(), Black);
     black_d3c5.print();
@@ -21,11 +22,12 @@ TEST_F(EvaluationTest, d3c5) {
     white_d3c5.translate_from_arr(TestBoard::d3c5(), White);
     white_d3c5.print();
 
-    EXPECT_EQ(0, evaluate(black_d3c5));
-    EXPECT_EQ(0, evaluate(white_d3c5));
+    EXPECT_EQ(0, evaluate(black_d3c5, a.cell_score));
+    EXPECT_EQ(0, evaluate(white_d3c5, a.cell_score));
 }
 
 TEST_F(EvaluationTest, d3c5e6) {
+    ArrStruct a = make_score();
     board black_d3c5e6;
     black_d3c5e6.translate_from_arr(TestBoard::d3c5e6(), Black);
     black_d3c5e6.print();
@@ -35,11 +37,12 @@ TEST_F(EvaluationTest, d3c5e6) {
     white_d3c5e6.translate_from_arr(TestBoard::d3c5e6(), White);
     white_d3c5e6.print();
 
-    EXPECT_EQ(-3, evaluate(black_d3c5e6));
-    EXPECT_EQ(3, evaluate(white_d3c5e6));
+    EXPECT_EQ(-3, evaluate(black_d3c5e6, a.cell_score));
+    EXPECT_EQ(3, evaluate(white_d3c5e6, a.cell_score));
 }
 
 TEST_F(EvaluationTest, all_black) {
+    ArrStruct a = make_score();
     board black_all_black;
     black_all_black.translate_from_arr(TestBoard::all_black(), Black);
     black_all_black.print();
@@ -48,6 +51,6 @@ TEST_F(EvaluationTest, all_black) {
     white_all_black.translate_from_arr(TestBoard::all_black(), White);
     white_all_black.print();
 
-    EXPECT_EQ(-104, evaluate(black_all_black));
-    EXPECT_EQ(104, evaluate(white_all_black));
+    EXPECT_EQ(-104, evaluate(black_all_black, a.cell_score));
+    EXPECT_EQ(104, evaluate(white_all_black, a.cell_score));
 }
